massive 1k line refactor. combined plg logic and env logic into 1 file, game now runs with bitmasking for  state management. 2x speed increase
