+++ pufferlib/ocean/tower_climb/tower_climb.h
@@ -5,8 +5,10 @@
 #include <string.h>
 #include "raylib.h"
 #include "raymath.h"
-#include "levels.h"
+// #include "levels.h"
 #include "rlgl.h"
+#include <time.h>
+
 #if defined(PLATFORM_DESKTOP)
     #define GLSL_VERSION            330
 #else   // PLATFORM_ANDROID, PLATFORM_WEB
@@ -24,21 +26,95 @@
 #define DOWN 1
 #define GRAB 4
 #define DROP 5
+// robot state
 #define DEFAULT 0
 #define HANGING 1
-#define HOLDING_BLOCK 2
-#define NUM_DIRECTIONS 4
-#define LEVEL_MAX_SIZE 1000
+// observation space
 #define PLAYER_OBS 4
 #define OBS_VISION 225
-#define MAX_FLOORS 9
-
-static const int DIRECTIONS[NUM_DIRECTIONS] = {0, 1, 2, 3};
-static const int DIRECTION_VECTORS_X[NUM_DIRECTIONS] = {1, 0, -1, 0};
-static const int DIRECTION_VECTORS_Z[NUM_DIRECTIONS] = {0, 1, 0, -1};
-
-
+// PLG VS ENV
+#define PLG_MODE 0
+#define RL_MODE 1
+//logs
 #define LOG_BUFFER_SIZE 1024
+// level size
+#define row_max 10
+#define col_max 10
+#define depth_max 10
+// block bytes
+#define BLOCK_BYTES 125
+// FNV-1a hash function
+#define FNV_OFFSET 0xcbf29ce484222325ULL
+#define FNV_PRIME  0x100000001b3ULL
+// moves
+#define MOVE_ILLEGAL 0
+#define MOVE_SUCCESS 1
+#define MOVE_DEATH 2
+// bitmask operations
+#define SET_BIT(mask, i)    ( (mask)[(i)/8] |=  (1 << ((i)%8)) )
+#define CLEAR_BIT(mask, i)  ( (mask)[(i)/8] &= ~(1 << ((i)%8)) )
+#define TEST_BIT(mask, i)   ( ((mask)[(i)/8] & (1 << ((i)%8))) != 0 )
+
+// BFS
+#define MAX_BFS_SIZE 70000000
+#define MAX_NEIGHBORS 6 // based on action space
+
+// hash table 
+#define TABLE_SIZE 70000003
+
+// actions
+#define ACTION_RIGHT 0
+#define ACTION_DOWN 1
+#define ACTION_LEFT 2
+#define ACTION_UP 3
+#define ACTION_GRAB 4
+#define ACTION_DROP 5
+
+// direction vectors
+#define NUM_DIRECTIONS 4
+static const int BFS_DIRECTION_VECTORS_X[NUM_DIRECTIONS] = {1, 0, -1, 0};
+static const int BFS_DIRECTION_VECTORS_Z[NUM_DIRECTIONS] = {0, 1, 0, -1};
+// shimmy wrap constants 
+static const int wrap_x[4][2] = {
+    {1,1},
+    {1,-1},
+    {-1,-1},
+    {-1, 1}
+};
+static const int wrap_z[4][2] = {
+    {-1, 1},
+    {1, 1},
+    {1, -1},
+    {-1, -1}
+};
+static const int wrap_orientation[4][2] = {
+    {ACTION_DOWN,ACTION_UP},
+    {ACTION_LEFT, ACTION_RIGHT},
+    {ACTION_UP, ACTION_DOWN},
+    {ACTION_RIGHT, ACTION_LEFT}
+};
+
+typedef struct Level Level;
+struct Level {
+    const int* map;
+    int rows;
+    int cols;
+    int size;
+    int total_length;
+    int goal_location;
+    int spawn_location;
+};
+static Level levels[3];  // Array to store the actual level objects
+
+
+typedef struct PuzzleState PuzzleState;
+struct PuzzleState {
+    unsigned char blocks[BLOCK_BYTES];
+    int robot_position;
+    int robot_orientation;
+    int robot_state;
+    int block_grabbed;
+};
 
 typedef struct Log Log;
 struct Log {
@@ -102,18 +178,8 @@ struct CTowerClimb {
     Log log;
     float score;
     int map_choice;
-    int robot_position; 
-    int robot_direction;
-    int robot_state;
-    int robot_orientation;
-    int* board_state; 
-    int* blocks_to_move;
-    int* blocks_to_fall;
-    int block_grabbed;
-    int rows_cleared;
-    int animation_state;
-    int previous_position;
     Level level;
+    PuzzleState state;  // Contains blocks bitmask, position, orientation, etc.
     int level_number;
     int distance_to_goal;
     float reward_climb_row;
@@ -123,63 +189,37 @@ struct CTowerClimb {
     float reward_distance;
 };
 
-float get_distance_to_goal(CTowerClimb* env) {
-	int sz = env->level.size;
-	int cols = env->level.cols;
-		    
-		    // Get robot position coordinates
-	int robot_floor = env->robot_position / sz;
-	int robot_grid = env->robot_position % sz;
-	int robot_x = robot_grid % cols;
-	int robot_z = robot_grid / cols;
-		    //                         // Get goal position coordinates using goal_location
-	int goal_floor = env->level.goal_location / sz;
-	int goal_grid = env->level.goal_location % sz;
-	int goal_x = goal_grid % cols;
-	int goal_z = goal_grid / cols;
-	// Calculate Manhattan distance
-	int x_dist = abs(robot_x - goal_x);
-	int z_dist = abs(robot_z - goal_z);
-	int floor_dist = abs(robot_floor - goal_floor);    
-	return x_dist + z_dist + floor_dist;
-}
+void levelToPuzzleState(const Level* level, PuzzleState* state) {
+    // 1) Clear the entire bitmask to 0
+    memset(state->blocks, 0, BLOCK_BYTES);
 
+    // 2) For each cell i in [0 .. total_length-1]:
+    //    if map[i] == 1 => set the bit in state->blocks
+    //    if map[i] == 0 or 2 => do nothing (they're empty or goal)
+    for (int i = 0; i < level->total_length; i++) {
+        if (level->map[i] == 1) {
+            SET_BIT(state->blocks, i);
+        }
+    }
 
-int get_direction(CTowerClimb* env, int action) {
-    // For reference: 
-    // 0 = right (initial), 1 = down, 2 = left, 3 = up
-    int current_direction = env->robot_orientation;
-    if (env->block_grabbed == -1){
-        return action;
-    }
-    // Check if direction is opposite (differs by 2) or perpendicular (differs by 1 or 3)
-    int diff = abs(current_direction - action);
-    if(diff == 1 || diff == 3){
-        env->block_grabbed = -1;
-    }
-    else if(diff == 2){
-        return current_direction;
-    }
-    return action;
+    // 3) Copy spawn location into the puzzle state
+    state->robot_position = level->spawn_location;
+
+    // 4) Initialize other fields as needed
+    state->robot_orientation = 3;  // e.g., "facing up" or whichever default
+    state->robot_state = 0;        // or define a known 'DEFAULT' constant
+    state->block_grabbed = -1;     // no block held initially
 }
 
 
 void init(CTowerClimb* env) {
     env->level_number = 0;
     env->level = levels[env->level_number];
-    env->board_state = (int*)calloc(LEVEL_MAX_SIZE, sizeof(int));
-    env->block_grabbed = -1;
-    env->blocks_to_move = (int*)calloc(10, sizeof(int));
-    env->blocks_to_fall = (int*)calloc(LEVEL_MAX_SIZE, sizeof(int));
-    env->rows_cleared = 0;
-    env->robot_orientation = UP;
-    env->robot_position = env->level.spawn_location;
-    env->distance_to_goal = get_distance_to_goal(env);
-    memcpy(env->board_state, env->level.map, env->level.total_length * sizeof(int));
-    memset(env->blocks_to_fall, -1, LEVEL_MAX_SIZE * sizeof(int));
-    for (int i = 0; i < env->level.cols; i++){
-        env->blocks_to_move[i] = -1;
-    }
+    
+    // Initialize PuzzleState
+    levelToPuzzleState(&env->level, &env->state);
+    printf("state: %d\n", env->state.robot_position);   
+    printf("level: %d\n", env->level.goal_location);
 }
 
 void allocate(CTowerClimb* env) {
@@ -192,9 +232,7 @@ void allocate(CTowerClimb* env) {
 }
 
 void free_initialized(CTowerClimb* env) {
-    free(env->blocks_to_move);
-    free(env->blocks_to_fall);
-    free(env->board_state);
+
 }
 
 void free_allocated(CTowerClimb* env) {
@@ -258,8 +296,8 @@ void compute_observations(CTowerClimb* env) {
     int max_floors = env->level.total_length / sz;
     
     // Get player position
-    int current_floor = env->robot_position / sz;
-    int grid_pos = env->robot_position % sz;
+    int current_floor = env->state.robot_position / sz;
+    int grid_pos = env->state.robot_position % sz;
     int player_x = grid_pos % cols;
     int player_z = grid_pos / cols;
     
@@ -332,13 +370,14 @@ void compute_observations(CTowerClimb* env) {
     }
 
     // Fill in observations
-    for (int y = 0; y < window_height; y++) {  // Always iterate through full window height
+    for (int y = 0; y < window_height; y++) {
         int world_y = y + y_start;
-        for (int z = 0; z < window_depth; z++) {  // Always iterate through full window depth
+        for (int z = 0; z < window_depth; z++) {
             int world_z = z + z_start;
-            for (int x = 0; x < window_width; x++) {  // Always iterate through full window width
+            for (int x = 0; x < window_width; x++) {
                 int world_x = x + x_start;
                 int obs_idx = x + z * window_width + y * (window_width * window_depth);
+                
                 // Check if position is out of bounds
                 int board_idx = world_y * sz + world_z * cols + world_x;
                 if (world_x < 0 || world_x >= cols || 
@@ -348,41 +387,33 @@ void compute_observations(CTowerClimb* env) {
                     env->observations[obs_idx] = -1.0f;
                     continue;
                 }
+                
                 // Position is in bounds, set observation
-                if (board_idx == env->robot_position) {
+                if (board_idx == env->state.robot_position) {
                     env->observations[obs_idx] = 3.0f;
                     continue;
                 }
-                env->observations[obs_idx] = (float)env->board_state[board_idx];
+                
+                // Use bitmask directly instead of board_state array
+                env->observations[obs_idx] = (float)TEST_BIT(env->state.blocks, board_idx);
             }
         }
     }
     
     // Add player state information at the end
     int state_start = window_width * window_depth * window_height;
-    env->observations[state_start] = (float)env->robot_orientation;
-    env->observations[state_start + 1] = (float)env->robot_state;
-    env->observations[state_start + 2] = (float)env->block_grabbed;
-    env->observations[state_start + 3] = (float)(env->block_grabbed != -1);
+    env->observations[state_start] = (float)env->state.robot_orientation;
+    env->observations[state_start + 1] = (float)env->state.robot_state;
+    env->observations[state_start + 2] = (float)env->state.block_grabbed;
+    env->observations[state_start + 3] = (float)(env->state.block_grabbed != -1);
 }
 
 void reset(CTowerClimb* env) {
     env->log = (Log){0};
     env->dones[0] = 0;
-    env->robot_orientation = UP;
-    env->robot_state = DEFAULT;
-    env->block_grabbed = -1;
-    env->rows_cleared = 0;
     env->level_number = 0;
     env->level = levels[env->level_number];
-    env->robot_position = env->level.spawn_location;
-    env->distance_to_goal = get_distance_to_goal(env);
-    memcpy(env->board_state, env->level.map, env->level.total_length * sizeof(int));
-    if(env->level.total_length < LEVEL_MAX_SIZE){
-	    memset(env->board_state + env->level.total_length, 0, (LEVEL_MAX_SIZE-env->level.total_length) * sizeof(int));
-    }
-    memset(env->blocks_to_move, -1, env->level.cols * sizeof(int));
-    memset(env->blocks_to_fall, -1, env->level.total_length * sizeof(int));
+    levelToPuzzleState(&env->level, &env->state);
     compute_observations(env);
 }
 
@@ -392,638 +423,449 @@ void illegal_move(CTowerClimb* env){
     env->log.episode_return += env->reward_illegal_move;
 }
 
-int get_local_direction(CTowerClimb* env, int action) {
-    // We assume action can be LEFT or RIGHT
-    // Orientation is one of (RIGHT=0, DOWN=1, LEFT=2, UP=3).
-    //  c mod robot orientation
-    if (action == LEFT) {
-        return (env->robot_orientation + 3) % 4;
-    } 
-    if (action == RIGHT) {
-        return (env->robot_orientation + 1) % 4;
-    }
-    return env->robot_orientation;
-}
-
-void handle_grab_block(CTowerClimb *env){
-    int sz = env->level.size;
-    int cols = env->level.cols;
-    int rows = env->level.rows;
-    int current_floor = env->robot_position / sz;
-    int grid_pos = env->robot_position % sz;
-    int x = grid_pos % cols;
-    int z = grid_pos / cols;
-
-    int dx = DIRECTION_VECTORS_X[env->robot_orientation]; // possibly make dx and dz 1D arrays
-    int dz = DIRECTION_VECTORS_Z[env->robot_orientation];
-
-    int next_x = x + dx;
-    int next_z = z + dz;
-    if (env->robot_state == HANGING){
-        illegal_move(env);
-        return;
-    }
-    if (next_x < 0 || next_x >= cols || next_z < 0 || next_z >= rows ) {
-        // Attempting to move outside the 4x4 grid on this floor - do nothing
-        return;
-    }
-    int next_index = sz*current_floor + cols*next_z + next_x;
-    int next_cell = env->board_state[next_index];
-    if (next_cell!=1){
-        illegal_move(env);
-        return;
-    }
-    if(env->block_grabbed == next_index){
-        env->block_grabbed = -1;
-    } else {
-        env->block_grabbed = next_index;
-    }
-    //env->rewards[0] = .2;
-    //env->log.episode_return += 0.2;
-}
-
-int is_next_to_block(CTowerClimb* env, int target_position){
-    int sz = env->level.size;
-    int cols = env->level.cols;
-    int current_floor = target_position / sz;
-    int grid_pos = target_position % sz;
-    int x = grid_pos % cols;
-    int z = grid_pos / cols;
-    int dx = DIRECTION_VECTORS_X[env->robot_orientation];
-    int dz = DIRECTION_VECTORS_Z[env->robot_orientation];
-    int next_x = x + dx;
-    int next_z = z + dz;    
-    int next_index = sz*current_floor + cols*next_z + next_x;  // semantically group multiplication
-    int next_cell = env->board_state[next_index];
-    if (next_cell == 1) {
-        return 1;
-    }
-    return 0;
-}
-
-void add_blocks_to_move(CTowerClimb* env, int interval){
-    int start_index = env->blocks_to_move[0];
-    int cols = env->level.cols;
-    for (int i = 0; i < cols; i++){
-        int b_address = start_index + i*interval;
-        if(i!=0 && env->blocks_to_move[i-1] == -1){
-            break;
-        }
-        if(env->board_state[b_address] == 1){
-            env->blocks_to_move[i] = b_address;
-        }
-    }
-}
-
-// Helper function to check if position is within bounds
-static inline int is_valid_position(int pos) {
-    return (pos >= 0 && pos < LEVEL_MAX_SIZE);
-}
-
-// Helper function to check block stability
-static int is_block_stable(CTowerClimb* env, int position) {
-    int sz = env->level.size;
-    int cols = env->level.cols;
-    // Check bottom support first
-    int bottom = position - sz;
-    if (is_valid_position(bottom) && env->board_state[bottom] == 1) {
-        return 1;
-    }
-
-    // Check edge supports only if no bottom support
-    // Left+Right support below
-    int left_support = position - sz - 1;
-    int right_support = position - sz + 1;
-    if (is_valid_position(left_support) && is_valid_position(right_support) &&
-        (env->board_state[left_support] == 1 || env->board_state[right_support] == 1)) {
-        return 1;
-    }
-
-    // Front+Back support below
-    int front_support = position - sz - cols;
-    int back_support = position - sz + cols;
-    if (is_valid_position(front_support) && is_valid_position(back_support) &&
-        (env->board_state[front_support] == 1 || env->board_state[back_support] == 1)) {
-        return 1;
-    }
-
-    return 0;
-}
-
-int  add_blocks_to_fall(CTowerClimb* env){
-    int sz = env->level.size;
-    int cols = env->level.cols;
-    int total_length = env->level.total_length;
-// Create queue as simple arrays with front/rear indices
-    int queue[total_length];  // Max possible blocks
-    int front = 0;
-    int rear = 0;
-    for(int i= 0;i< total_length;i++){
-        if(env->blocks_to_fall[i] == -1){
-            break;
-        }
-        queue[rear] = env->blocks_to_fall[i];
-        rear++;
-    }
-    // Helper to add block to queue if it's a valid block
-    for(int i = 0; i < env->level.cols; i++) {
-        // invert conditional
-        if(env->blocks_to_move[i] == -1){
-            continue;
-        }
-        // Add block directly above
-        int cell_above = env->blocks_to_move[i] + sz;
-        
-        // If valid block above and unstable, add to queue
-        if (cell_above < total_length && (env->board_state[cell_above] == 1 || env->board_state[cell_above] ==2)
-        && !is_block_stable(env, cell_above)) {
-            queue[rear++] = cell_above;
-        }
-        
-        // Check edge-supported blocks
-        int edge_blocks[4] = {
-            cell_above - 1,  // left
-            cell_above + 1,  // right
-            cell_above - cols,  // front
-            cell_above + cols   // back
-        };
-        
-        // Add valid edge blocks to queue
-        for(int j = 0; j < 4; j++) {
-            if (edge_blocks[j] >= 0 && edge_blocks[j] < total_length && 
-                (env->board_state[edge_blocks[j]] == 1 || env->board_state[edge_blocks[j]] ==2)
-                && !is_block_stable(env, edge_blocks[j])) {
-                queue[rear++] = edge_blocks[j];
-            }
-        }
-    }
-
-    // Process queue until empty
-    while(front < rear) {
-        int current = queue[front++];
-        int falling_position = current;
-        int found_support = 0;
-        if(env->board_state[current]==2){
-            return 0;
-        }
-        // Remove block from current position
-        env->board_state[current] = 0;
-        // Keep moving down until support found or bottom reached
-        while (!found_support && falling_position >= sz) {  // Check if we still have space below            
-            // Place block temporarily to check stability
-            env->board_state[falling_position] = 1;
-            if (is_block_stable(env, falling_position)) {
-                found_support = 1;
-            } else {
-                // Remove block if not stable
-                env->board_state[falling_position] = 0;
-            }
-            if (!found_support) {
-                falling_position -= sz;  // Move down one level
-            }
-        }
-          
-        // Add blocks that might be affected by this fall
-        int original_above = current + sz;  // Block directly above original position
-        if (original_above < total_length && (env->board_state[original_above] == 1 || env->board_state[original_above] ==2) && !is_block_stable(env, original_above)) {
-            queue[rear++] = original_above;
-        }
-        
-        int edge_check[4] = {
-            original_above - 1,  // Left of original
-            original_above + 1,  // Right of original
-            original_above - cols,  // Front of original
-            original_above + cols   // Back of original
-        };
-        
-        for(int i = 0; i < 4; i++) {
-            if (edge_check[i] >= 0 && edge_check[i] < total_length && 
-                (env->board_state[edge_check[i]] == 1 || env->board_state[edge_check[i]] ==2)
-                && !is_block_stable(env, edge_check[i])) {
-                queue[rear++] = edge_check[i];
-            }
-        }
-        
-        // If no support found, block disappears (already set to 0)
-    }
-    // death by block crush
-    memset(env->blocks_to_fall, -1, LEVEL_MAX_SIZE * sizeof(int));
-    memset(env->blocks_to_move, -1, cols * sizeof(int));
-    return 1;
-}
-void move_blocks(CTowerClimb* env, int interval){
-    int count = 0;
-    int cols = env->level.cols;
-    int rows = env->level.rows;
-    int sz = env->level.size;
-    // cache reused variables 
-    for (int i = 0; i < cols; i++){
-        // invert conditional 
-        int b_index = env->blocks_to_move[i];
-        if (b_index == -1){
-            continue;
-        }
-        if(i==0){
-            env->board_state[b_index] = 0;
-        }
-        int grid_pos = b_index % sz;
-        int x = grid_pos % cols;
-        int z = grid_pos / cols;
-        // Check if movement would cross floor boundaries
-        if ((x == 0 && interval == -1) ||           // Don't move left off floor
-            (x == cols-1 && interval == 1) ||       // Don't move right off floor
-            (z == 0 && interval == -cols) ||        // Don't move forward off floor
-            (z == rows-1 && interval == cols)) {    // Don't move back off floor
-            return;
-        }
-
-        // If we get here, movement is safe
-        env->board_state[b_index + interval] = 1;
-        env->blocks_to_fall[count] = b_index + interval;
-        count++;
-    }
-}
-
-void shimmy(CTowerClimb* env, int action, int current_floor, int x, int z, int x_mod, int z_mod, int x_direction_mod, int z_direction_mod, int final_orienation){
-    int sz = env->level.size;
-    int cols = env->level.cols;
-    int next_block = sz*current_floor + (z+z_mod)*cols + (x+x_mod);
-    int destination_block = sz*current_floor + (z+z_direction_mod)*cols + (x+x_direction_mod);
-    if (env->board_state[next_block] == 0){
-        illegal_move(env);
-        return;
-    }
-    int above_destination = destination_block + sz;
-    // cant wrap shimmy with block above.
-    if(env->board_state[above_destination] == 1){
-        illegal_move(env);
-        return;
-    }
-    env->robot_position = destination_block;
-    env->robot_orientation = final_orienation;
-    return;
-}
-
-void handle_climb(CTowerClimb* env, int action, int current_floor,int x, int z, int next_z, int next_x, int next_index, int next_cell){
-    if (!(next_cell == 1 || next_cell == 2) || env->block_grabbed != -1) {
-        illegal_move(env);
-        return;
-    }
-    if (current_floor >= MAX_FLOORS){
-        return;
-    }
-    int sz = env->level.size;
-    int cols = env->level.cols;
-    int climb_index = sz*(current_floor + 1) + cols*next_z + next_x;
-    int climb_cell = env->board_state[climb_index];
-    int direct_above = env->robot_position + sz;
-    int direct_above_cell = env->board_state[direct_above];
-    if ((climb_cell != 0 || direct_above_cell != 0) && next_cell != 2){
-        illegal_move(env);
-        return;
-    }
-    env->robot_position = climb_index;
-    env->robot_state = DEFAULT; // set hanging to indicate we climbed a block
-    int previous_row = (env->robot_position - sz) / sz;
-    if(previous_row <= env->rows_cleared){
-        return;
-    }
-    env->rows_cleared = previous_row;
-    env->rewards[0] = env->reward_climb_row;
-    env->log.episode_return += env->reward_climb_row;  
-}
-
-
-void handle_down(CTowerClimb* env, int action, int current_floor,int x, int z, int next_z, int next_x, int next_index, int next_cell){
-    int sz = env->level.size;
-    int cols = env->level.cols;
-    int below_index = sz*(current_floor - 1) + cols*next_z + next_x;
-    int below_cell;
-    if(below_index >= 0){
-	   below_cell =  env->board_state[below_index];
-    } else {
-	    below_cell = -1;
-    }
-    int below_next_index = below_index - sz;
-    int below_next_cell;
-    if (below_next_index >= 0){
-	    below_next_cell = env->board_state[below_next_index];
-    } else{
-    	below_next_cell = 0;
-    }
-    // Default state cases
-    if (below_cell == 1 && env->robot_state == DEFAULT) {
-        env->robot_position = next_index;
-        env->robot_state = DEFAULT;
-        return;
-    }
-    if (below_cell == 0 && below_next_cell == 0 && env->robot_state == DEFAULT) {
-        env->robot_position = below_index;
-        env->robot_state = HANGING;
-        //mod ti thing
-        if(env->robot_direction == RIGHT){
-            env->robot_orientation = LEFT;
-        }
-        if(env->robot_direction == LEFT){
-            env->robot_orientation = RIGHT;
-        }
-        if(env->robot_direction == DOWN){
-            env->robot_orientation = UP;
-        }
-        if(env->robot_direction == UP){
-            env->robot_orientation = DOWN;
-        }
-        return;
-    }
-    if (below_cell == 0 && below_next_cell == 1 && env->robot_state == DEFAULT){
-        env->robot_position = below_index;
-        env->robot_state = DEFAULT;
-        env->rewards[0] = env->reward_fall_row;
-        env->log.episode_return += env->reward_fall_row;
-    }
-}
-
-void handle_left_right(CTowerClimb* env, int action, int current_floor,int x, int z, int next_z, int next_x, int next_index, int next_cell){
-    // Check if the cell in front is free, the goal, or blocked
-    int sz = env->level.size;
-    int cols = env->level.cols;
-    int rows = env->level.rows;
-    int below_index = sz*(current_floor - 1) + cols*next_z + next_x;
-    int below_cell;
-    if (below_index >=0){
-	below_cell = env->board_state[below_index];
-    } else {
-	    below_cell = -1;
-    }
-    // shimmying
-    if (env->robot_state == HANGING){
-        int local_direction = get_local_direction(env, action);
-        int local_next_dx = DIRECTION_VECTORS_X[local_direction];
-        int local_next_dz = DIRECTION_VECTORS_Z[local_direction];
-        int local_next_x = x + local_next_dx;
-        int local_next_z = z + local_next_dz;
-        if(local_next_x < 0 || local_next_z < 0 || local_next_x >= cols || local_next_z >= rows){
-            illegal_move(env);
-            return;
-        }
-        int local_next_index = sz*current_floor + cols*local_next_z + local_next_x;
-        int local_next_cell = env->board_state[local_next_index];
-        if (is_next_to_block(env, local_next_index) && local_next_cell == 0){
-            // standard shimmy
-            int above_index = local_next_index + sz;
-            int above_cell = env->board_state[above_index];
-            // cant shimmy with block above
-            if (above_cell == 1 ){
-                illegal_move(env);
-                return;
-            }
-            env->robot_position = local_next_index;
-            return;
-        }
-
-        if(local_next_cell == 1){
-            static const int LEFT_TURNS[] = {3, 0, 1, 2};   // RIGHT->UP, DOWN->RIGHT, LEFT->DOWN, UP->LEFT
-            static const int RIGHT_TURNS[] = {1, 2, 3, 0};  // RIGHT->DOWN, DOWN->LEFT, LEFT->UP, UP->RIGHT
-
-            env->robot_orientation = (action == LEFT) ? 
-                LEFT_TURNS[env->robot_orientation] : 
-                RIGHT_TURNS[env->robot_orientation];
-            return;
-        }
-        
-        int current_floor = env->robot_position / sz;
-        
-        if (env->robot_orientation == UP) {  // Hanging on north face
-            if (action == RIGHT) {
-                shimmy(env, action, current_floor, x, z, 0, -1, 1, -1, LEFT);
-                env->robot_direction = RIGHT;   
-            }
-            else if (action == LEFT) {
-                shimmy(env, action, current_floor, x, z, 0,-1, -1, -1, RIGHT);
-                env->robot_direction = LEFT;
-            }
-        }
-        // Similar logic for other directions (DOWN, LEFT, RIGHT)
-        else if (env->robot_orientation == DOWN) {
-            if (action == RIGHT) {
-                shimmy(env, action, current_floor, x, z, 0, 1, -1, 1, RIGHT);
-                env->robot_direction = LEFT;
-            }
-            if (action == LEFT) {
-                shimmy(env, action, current_floor, x, z, 0, 1, 1, 1, LEFT);
-                env->robot_direction = RIGHT;   
-            }
-        }
-        else if (env->robot_orientation == RIGHT) {
-            if (action == LEFT) {
-                // Check block to our right
-                shimmy(env, action, current_floor, x, z, 1, 0, 1, -1, DOWN);
-                env->robot_direction = LEFT;
-            }
-            else if (action == RIGHT) {
-                // Check block to our right
-                shimmy(env, action, current_floor, x, z, 1, 0, 1, 1, UP);
-                env->robot_direction = RIGHT;
-            }
-        }
-        else if (env->robot_orientation == LEFT) {
-            if (action == RIGHT) {
-                shimmy(env,action, current_floor, x, z, -1, 0, -1, -1, DOWN);
-                env->robot_direction = RIGHT;
-            }
-            if (action == LEFT) {
-                shimmy(env,action, current_floor, x, z, -1, 0, -1, 1, UP);
-                env->robot_direction = LEFT;
-            }
-        }
-    }
-
-    if ((next_cell ==1 || next_cell ==2) && env->robot_state == DEFAULT){
-        handle_climb(env, action, current_floor, x, z, next_z, next_x, next_index, next_cell);
-    }
-    if (next_cell == 0 && below_cell == 1 && env->robot_state == DEFAULT){
-        env->robot_position = next_index;
-        env->robot_state = DEFAULT;
-        env->robot_orientation = env->robot_direction;
-    }
-    // dropping to hang state
-    if (next_cell == 0 && below_cell == 0 && env->robot_state == DEFAULT){
-        handle_down(env, action, current_floor, x, z, next_z, next_x, next_index, next_cell);
-    }   
-}
-void handle_unstable_goal(CTowerClimb* env){
+void death(CTowerClimb* env){
 	env->rewards[0] = -1;
 	env->log.episode_return -= 1;
 	add_log(env->log_buffer, &env->log);
 	reset(env);
 }
 
-void handle_move_forward(CTowerClimb* env, int action) {
-    int sz = env->level.size;
-    int cols = env->level.cols;
-    int rows = env->level.rows;
-    // Calculate current floor, x, z from the robot_position
-    int current_floor = env->robot_position / sz;
-    int grid_pos = env->robot_position % sz;
-    int x = grid_pos % cols;
-    int z = grid_pos / cols;
 
-    // Determine the offset for the next cell based on env->robot_direction
-    // make this 1D
-    int dx = DIRECTION_VECTORS_X[env->robot_direction];
-    int dz = DIRECTION_VECTORS_Z[env->robot_direction];
-    
-    int orient_dx = DIRECTION_VECTORS_X[env->robot_orientation];
-    int orient_dz = DIRECTION_VECTORS_Z[env->robot_orientation];
+int isGoal(const PuzzleState* s, const Level* lvl) {
+    // 1) Check if player is at the goal cell
+    if (s->robot_position - 100 != lvl->goal_location) return 0;
 
-    int front_x = x + orient_dx;
-    int front_z = z + orient_dz;
-    int front_index = sz*current_floor + cols*front_z + front_x;
-    int front_cell = env->board_state[front_index];
-    int front_below_index = front_index - sz;
-    int front_below_cell;
-    if(front_below_index >= 0){
-	    front_below_cell = env->board_state[front_below_index];
+    return 1;  // all conditions satisfied
+}
+
+int move(PuzzleState* outState, int action, int mode, CTowerClimb* env, const Level* lvl){
+    int new_position = outState->robot_position + BFS_DIRECTION_VECTORS_X[action] + BFS_DIRECTION_VECTORS_Z[action]*10;
+    outState->robot_position = new_position;
+    return 1;
+}
+
+int climb(PuzzleState* outState, int action, int mode, CTowerClimb* env, const Level* lvl){
+    int cell_direct_above = outState->robot_position + lvl->size;
+    int cell_next_above = cell_direct_above + BFS_DIRECTION_VECTORS_X[action] + BFS_DIRECTION_VECTORS_Z[action]*lvl->cols;
+    int goal = lvl->goal_location;
+    int can_climb = cell_direct_above < lvl->total_length && cell_next_above < lvl->total_length &&
+     !TEST_BIT(outState->blocks, cell_direct_above) &&
+      !TEST_BIT(outState->blocks, cell_next_above) && 
+      cell_next_above != goal &&
+      cell_direct_above != goal;
+    if (can_climb){
+        outState->robot_position = cell_next_above;
+        outState->robot_state = 0;
+        return 1;
+    }
+    return 0;
+}
+
+int drop(PuzzleState* outState, int action, int mode, CTowerClimb* env, const Level* lvl){
+    int next_cell = outState->robot_position + BFS_DIRECTION_VECTORS_X[action] + BFS_DIRECTION_VECTORS_Z[action]*lvl->cols;
+    int next_below_cell = next_cell - lvl->size;
+    int next_double_below_cell = next_cell - 2*lvl->size;
+    if (next_below_cell < 0){
+        return 0;
+    } 
+    int step_down = next_double_below_cell >= 0 && TEST_BIT(outState->blocks, next_double_below_cell);
+    if (step_down){
+        outState->robot_position = next_below_cell;
+        return 1;
     } else {
-	    front_below_cell = -1;
+        outState->robot_position = next_below_cell;
+        outState->robot_orientation = (outState->robot_orientation + 2) % 4;
+        outState->robot_state = 1;
+        return 1;
     }
-    int front_below_below_index = front_below_index - sz;
-    int front_below_below_cell;
-    if(front_below_below_index>=0){
-        front_below_below_cell = env->board_state[front_below_below_index];
-    } else {
-        front_below_below_cell = -1;
-    }
-    // Calculate the next potential cell's x, z
-    int next_x = x + dx;
-    int next_z = z + dz;
-    if(env->block_grabbed != -1 && abs(env->robot_direction - action) == 2){
-        // inverse direction based on direction vectors 
-        dx = DIRECTION_VECTORS_X[action];
-        dz = DIRECTION_VECTORS_Z[action];
-        next_x = x + dx;
-        next_z = z + dz;
-    }
+}
 
-    if (front_x < 0 || front_z < 0 || front_x >= cols || front_z >= rows){
-        illegal_move(env);
-        return;
+int drop_from_hang(PuzzleState* outState, int action, int mode, const Level* lvl){
+    int below_cell = outState->robot_position - lvl->size;
+    while(below_cell > lvl->size && !TEST_BIT(outState->blocks, below_cell)){
+        below_cell -= lvl->size;
     }
+    if (below_cell < lvl->size){
+        if (mode == PLG_MODE){
+            return MOVE_ILLEGAL;
+        }
+        if (mode == RL_MODE){
+            return MOVE_DEATH;
+        }
+    }
+    outState->robot_position = below_cell+lvl->size;
+    outState->robot_state = 0;
+    return 1;
+}
+static inline int bfs_is_valid_position(int pos, const Level* level) {
+    return (pos >= 0 && pos < level->total_length);
+}
+
+// Helper function to check block stability
+static int bfs_is_block_stable(const PuzzleState* state, int position, const Level* level) {
+    const int fs = level->size;
+    const int positions[] = {
+        position - fs,              // Bottom
+        position - fs - 1,          // Left
+        position - fs + 1,          // Right
+        position - fs - level->cols, // Front
+        position - fs + level->cols  // Back
+    };
     
-    // Convert next x,z to a linear index for the same floor
-    int next_index = sz*current_floor + cols*next_z + next_x;
-    int next_cell = env->board_state[next_index];
-    if((action == LEFT || action == RIGHT) && env->block_grabbed == -1){
-        handle_left_right(env, action, current_floor, x, z, next_z, next_x, next_index, next_cell);
+    for (int i = 0; i < 5; i++) {
+        if (bfs_is_valid_position(positions[i], level) && TEST_BIT(state->blocks, positions[i])) {
+            return 1;
+        }
+    }
+    return 0;
+}
+
+int will_fall(PuzzleState* outState, int position, const Level* lvl){
+    return bfs_is_valid_position(position, lvl) &&
+     (TEST_BIT(outState->blocks, position) || position == lvl->goal_location)
+      && !bfs_is_block_stable(outState, position, lvl);
+}
+
+int handle_block_falling(PuzzleState* outState, int* affected_blocks, int* blocks_to_move, int affected_blocks_count, const Level* lvl) {
+    // Create queue for blocks that need to be checked for falling
+    int bfs_queue[lvl->total_length];  // Max possible blocks to check
+    int front = 0;
+    int rear = 0;
+    int fs = lvl->size;
+    int cols = lvl->cols;
+
+    // Add initially affected blocks to queue
+    for (int i = 0; i < affected_blocks_count; i++) {
+        if (affected_blocks[i] == -1){
+            break;
+        }
+        bfs_queue[rear] = affected_blocks[i];
+        rear++;
     }
 
-    if (action == UP && env->block_grabbed == -1){
-        if(front_cell ==1 || front_cell ==2 ){
-            handle_climb(env, action, current_floor, x, z, front_z, front_x, front_index, front_cell);
+    // First check all blocks above and adjacent to moved blocks
+    for (int i = 0; i < lvl->cols; i++) {
+        int block_pos = blocks_to_move[i];
+        if (block_pos == -1){
+            continue;
         }
-        if(front_cell == 0 && front_below_cell == 1){
-            env->robot_position = front_index;
-            env->robot_state = DEFAULT;
-            env->robot_orientation = UP;
+        // Add block directly above
+        int cell_above = block_pos + fs;  // Assuming 100 is floor height
+        
+        // If valid block above and unstable, add to queue
+        if (will_fall(outState, cell_above, lvl)) {
+            bfs_queue[rear++] = cell_above;
         }
-        if(front_cell == 0 && front_below_cell == 0 && front_below_below_cell == 0){
-            env->robot_position = front_below_index;
-            env->robot_state = HANGING;
-            env->robot_orientation = DOWN;
-        }
-        if(front_cell == 0 && front_below_cell == 0 && front_below_below_cell == 1){
-            env->robot_position = front_below_index;
-            env->robot_state = DEFAULT;
+        
+        // Check edge-supported blocks
+        int edge_blocks[4] = {
+            cell_above - 1,      // left
+            cell_above + 1,      // right
+            cell_above - cols,     // front (assuming 10 is width)
+            cell_above + cols      // back
+        };
+        
+        // Add valid edge blocks to queue
+        for (int j = 0; j < 4; j++) {
+            if (will_fall(outState, edge_blocks[j], lvl)) {
+                bfs_queue[rear++] = edge_blocks[j];
+            }
         }
     }
-    if (action == DOWN && env->block_grabbed == -1){
-        if (env->robot_state == HANGING){
-            illegal_move(env);
-            return;
+
+    // Process queue until empty
+    while (front < rear) {
+        int current = bfs_queue[front++];
+        int falling_position = current;
+        int found_support = 0;
+
+        // Check if block is goal (2)
+        if (current == lvl->goal_location) {
+            // Goal block is falling - level failed
+            return 0;
         }
-        if (front_cell == 1 || front_cell == 2){
-            handle_climb(env, action, current_floor, x, z, front_z, front_x, front_index, front_cell);
-        } else {
-            handle_down(env, action, current_floor, x, z, next_z, next_x, next_index, next_cell);
-        }
-    }
-    if (front_cell == 1 && env->block_grabbed != -1) {
-        env->blocks_to_move[0] = front_index;
-        // Calculate block position based on direction
-        int block_offset = (env->robot_direction == 3) ? -cols :  // North
-                          (env->robot_direction == 1) ? cols :    // South
-                          (env->robot_direction == 0) ? 1 :    // East
-                          (env->robot_direction == 2) ? -1 :   // West
-                          -cols;  // Default to north behavior for other direction
-        // Pushing
-        if(abs(env->robot_orientation - action) == 0){
-            add_blocks_to_move(env, block_offset);
-            move_blocks(env, block_offset);
-            int stable = add_blocks_to_fall(env);
-            if (stable == 0){
-                handle_unstable_goal(env);
-            }
-            //env->rewards[0] = 0.25;
-            //env->log.episode_return += 0.25;
-        }
-        // Pulling
-        if (abs(env->robot_orientation - action) == 2){
-            int below_index = sz*(current_floor - 1) + cols*next_z + next_x;
-            int below_cell = env->board_state[below_index];
-            if(next_cell == 1){
-                illegal_move(env);
-                return;
-            }
-            env->blocks_to_move[0] = front_index;
-            env->board_state[front_index] = 0;
-            env->board_state[next_index + block_offset] = 1;
-            env->block_grabbed = next_index + block_offset;
-            int stable = add_blocks_to_fall(env);
-            if(stable ==0){
-		    handle_unstable_goal(env);
-		    return;
-	    }
-	    env->blocks_to_move[0] = -1;
-            if (below_cell == 0 && env->robot_state == DEFAULT){
-                env->robot_position = below_index;
-                env->robot_state = HANGING;
+
+        // Remove block from current position
+        CLEAR_BIT(outState->blocks, current);
+
+        // Keep moving down until support found or bottom reached
+        while (!found_support && falling_position >= fs) {  // 100 represents one floor down
+            // Place block temporarily to check stability
+            SET_BIT(outState->blocks, falling_position);            
+            // Check if block is stable
+            if (bfs_is_block_stable(outState, falling_position, lvl)) {
+                found_support = 1;
             } else {
-                env->robot_position = next_index;
+                // Remove block if not stable
+                CLEAR_BIT(outState->blocks, falling_position);
+                falling_position -= fs;  // Move down one level
             }
-            //env->rewards[0] = 0.25;
-            //env->log.episode_return +=0.25;
         }
-        if (env->robot_position != next_index){
-            env->block_grabbed = -1;
+
+        // Check blocks that might be affected by this fall
+        int original_above = current + fs;  // Block directly above original position
+        if (will_fall(outState, original_above, lvl)) {
+            bfs_queue[rear++] = original_above;
         }
-        memset(env->blocks_to_move, -1, cols * sizeof(int));
+
+        // Check edge blocks that might be affected
+        int edge_blocks[4] = {
+            original_above - 1,   // Left
+            original_above + 1,   // Right
+            original_above - cols,  // Front (assuming 10 is the width)
+            original_above + cols   // Back
+        };
+
+        for (int i = 0; i < 4; i++) {
+            if (will_fall(outState, edge_blocks[i], lvl)) {
+                bfs_queue[rear++] = edge_blocks[i];
+            }
+        }
     }
+    return 1;
 }
 
-void handle_drop(CTowerClimb* env){
-    env->robot_state = DEFAULT;
-    int sz = env->level.size;
-    int next_position = env->robot_position - sz;
-    // fallen off the map
-    if (env->robot_position < sz){
-        env->rewards[0] = -1;
-        env->log.episode_return -= 1;
-        add_log(env->log_buffer, &env->log);
-        reset(env);
-        return;
+int push(PuzzleState* outState, int action, const Level* lvl, int mode, CTowerClimb* env){
+    int first_block_index = outState->robot_position + BFS_DIRECTION_VECTORS_X[outState->robot_orientation] + BFS_DIRECTION_VECTORS_Z[outState->robot_orientation]*lvl->cols;
+    int block_offset = (outState->robot_orientation == 3) ? -lvl->cols :  // North
+                          (outState->robot_orientation == 1) ? lvl->cols :    // South
+                          (outState->robot_orientation == 0) ? 1 :    // East
+                          (outState->robot_orientation == 2) ? -1 :   // West
+                          -lvl->cols;
+                          
+    int* blocks_to_move = calloc(lvl->cols, sizeof(int));
+    for(int i = 0; i < lvl->cols; i++) {
+        blocks_to_move[i] = (i == 0) ? first_block_index : -1;
     }
-    while (env->board_state[next_position] != 1){
-        if(next_position < sz){
-            env->rewards[0] = -1;
-            env->log.episode_return -= 1;
-            add_log(env->log_buffer, &env->log);
-            reset(env);
-            return;
-        } 
-        next_position = next_position - sz;  
+    for (int i = 0; i < lvl->cols; i++){
+        int b_address = first_block_index + i*block_offset;
+        if(i!=0 && blocks_to_move[i-1] == -1){
+            break;
+        }
+        if(TEST_BIT(outState->blocks, b_address)){
+            blocks_to_move[i] = b_address;
+        }
     }
-    env->robot_position = next_position + sz;
+    int affected_blocks[lvl->cols];
+    int count = 0;
+    for (int i = 0; i < lvl->cols; i++){
+        // invert conditional 
+        int b_index = blocks_to_move[i];
+        if (b_index == -1){
+            continue;
+        }
+        if(i==0){
+            CLEAR_BIT(outState->blocks, b_index);
+        }
+        int grid_pos = b_index % lvl->size;
+        int x = grid_pos % lvl->cols;
+        int z = grid_pos / lvl->cols;
+        // Check if movement would cross floor boundaries
+        if ((x == 0 && block_offset == -1) ||           // Don't move left off floor
+            (x == lvl->cols-1 && block_offset == 1) ||       // Don't move right off floor
+            (z == 0 && block_offset == -lvl->cols) ||        // Don't move forward off floor
+            (z == lvl->rows-1 && block_offset == lvl->cols)) {    // Don't move back off floor
+            continue;
+        }
+
+        // If we get here, movement is safe
+        SET_BIT(outState->blocks, b_index + block_offset);
+        affected_blocks[count] = b_index + block_offset;
+        count++;
+    }
+    outState->block_grabbed = -1;
+    return handle_block_falling(outState, affected_blocks, blocks_to_move,count, lvl);
 }
+
+int pull(PuzzleState* outState, int action, const Level* lvl, int mode, CTowerClimb* env){
+    int pull_block = outState->robot_position + BFS_DIRECTION_VECTORS_X[outState->robot_orientation] + BFS_DIRECTION_VECTORS_Z[outState->robot_orientation]*10;
+    int block_offset = (action == 3) ? -lvl->cols :  // North
+                          (action == 1) ? lvl->cols :    // South
+                          (action == 0) ? 1 :    // East
+                          (action == 2) ? -1 :   // West
+                          -lvl->cols;
+    int block_in_front = TEST_BIT(outState->blocks, pull_block);
+    int cell_below_next_position = outState->robot_position + block_offset - lvl->size;
+    int backwards_walkable = bfs_is_valid_position(cell_below_next_position, lvl) && TEST_BIT(outState->blocks, cell_below_next_position);
+    if (block_in_front){
+        CLEAR_BIT(outState->blocks, pull_block);
+        SET_BIT(outState->blocks, outState->robot_position);
+        if (backwards_walkable){
+            outState->block_grabbed = outState->robot_position;
+            outState->robot_position = outState->robot_position + block_offset;
+        }
+        else {
+            outState->robot_position = cell_below_next_position;
+            outState->robot_state = 1;
+            outState->block_grabbed = -1;
+        }
+    }
+    int blocks_to_move[1];
+    blocks_to_move[0] = pull_block;
+    int affected_blocks[1] = {-1};
+    return handle_block_falling(outState, affected_blocks, blocks_to_move, 1, lvl);
+}
+
+int shimmy_normal(PuzzleState* outState, int action, const Level* lvl, int local_direction, int mode, CTowerClimb* env){
+    int next_cell = outState->robot_position + BFS_DIRECTION_VECTORS_X[local_direction] + BFS_DIRECTION_VECTORS_Z[local_direction]*lvl->cols;
+    int above_next_cell = next_cell + lvl->size;
+    if (bfs_is_valid_position(above_next_cell, lvl) && !TEST_BIT(outState->blocks, above_next_cell)){
+        outState->robot_position = next_cell;
+        return 1;
+    }
+    return 0;
+}
+
+int wrap_around(PuzzleState* outState, int action, const Level* lvl, int mode, CTowerClimb* env){
+    int action_idx = (action == ACTION_LEFT) ? 0 : 1;
+    int grid_pos = outState->robot_position % lvl->size;
+    int x = grid_pos % lvl->cols;
+    int z = grid_pos / lvl->cols;
+    int current_floor = outState->robot_position / lvl->size;
+    int new_x = x + wrap_x[outState->robot_orientation][action_idx];
+    int new_z = z + wrap_z[outState->robot_orientation][action_idx];
+    int new_pos = new_x + new_z*lvl->cols + current_floor*lvl->size;
+    if (TEST_BIT(outState->blocks, new_pos + lvl->size)){
+        return 0;
+    }
+    outState->robot_position = new_pos;
+    outState->robot_orientation = wrap_orientation[outState->robot_orientation][action_idx];
+    return 1;
+}
+
+int climb_from_hang(PuzzleState* outState, int action, const Level* lvl, int next_cell, int mode, CTowerClimb* env){
+    int climb_index = next_cell + lvl->size;
+    int direct_above_index = outState->robot_position + lvl->size;
+    int can_climb = bfs_is_valid_position(climb_index, lvl) && bfs_is_valid_position(direct_above_index, lvl) 
+    && !TEST_BIT(outState->blocks, climb_index) && !TEST_BIT(outState->blocks, direct_above_index);
+    if (can_climb){
+        outState->robot_position = climb_index;
+        outState->robot_state = 0;
+        return 1;
+    }
+    return 0;
+}
+
+int applyAction(PuzzleState* outState, int action, const Level* lvl, int mode, CTowerClimb* env) {
+    // necessary variables
+    int next_dx = BFS_DIRECTION_VECTORS_X[outState->robot_orientation];
+    int next_dz = BFS_DIRECTION_VECTORS_Z[outState->robot_orientation];   
+    int grid_pos = outState->robot_position % lvl->size;
+    int x = grid_pos % lvl->cols;
+    int z = grid_pos / lvl->cols;
+    int next_cell = outState->robot_position + next_dx + next_dz*lvl->cols;
+    int next_below_cell = next_cell - lvl->size;
+    int walkable = bfs_is_valid_position(next_below_cell, lvl) && TEST_BIT(outState->blocks, next_below_cell);
+    int block_in_front = bfs_is_valid_position(next_cell, lvl) && (TEST_BIT(outState->blocks, next_cell) || next_cell == lvl->goal_location);
+    int movement_action = (action >= 0 && action < 4);
+    int move_orient_check = (action == outState->robot_orientation);
+    int standing_and_holding_nothing = outState->robot_state == 0 && outState->block_grabbed == -1;
+    int hanging = outState->robot_state == 1;
+    // Handle movement actions with common orientation check
+    if (standing_and_holding_nothing && movement_action) {
+        // If orientation doesn't match action, just rotate
+        if (!move_orient_check) {
+            outState->robot_orientation = action;
+            return 1;
+        }
+        // Now handle the actual movement cases
+        if (walkable && !block_in_front) {
+            return move(outState, action, mode, env, lvl);
+        }
+        if (block_in_front) {
+            return climb(outState, action, mode, env, lvl);
+        }
+        if (!block_in_front && !walkable) {
+            return drop(outState, action, mode, env, lvl);
+        }
+    }
+
+    if(hanging && movement_action){
+        if (action == ACTION_UP){
+            return climb_from_hang(outState, action, lvl, next_cell, mode, env);
+        }
+        if (action == ACTION_DOWN){
+            return 0;
+        }
+        int local_direction = outState->robot_orientation;
+        if (action == ACTION_LEFT){
+            local_direction = (outState->robot_orientation + 3) % 4;
+        }
+        if (action == ACTION_RIGHT){
+            local_direction = (outState->robot_orientation + 1) % 4;
+        }
+        int shimmy_cell = next_cell + BFS_DIRECTION_VECTORS_X[local_direction] + BFS_DIRECTION_VECTORS_Z[local_direction]*lvl->cols;
+        int shimmy_path_cell = outState->robot_position + BFS_DIRECTION_VECTORS_X[local_direction] + BFS_DIRECTION_VECTORS_Z[local_direction]*lvl->cols;
+        int basic_shimmy = bfs_is_valid_position(shimmy_cell, lvl) && bfs_is_valid_position(shimmy_path_cell, lvl) && TEST_BIT(outState->blocks, shimmy_cell) && !TEST_BIT(outState->blocks, shimmy_path_cell);
+        int rotation_shimmy = bfs_is_valid_position(shimmy_path_cell, lvl) && TEST_BIT(outState->blocks, shimmy_path_cell);
+        int in_bounds = x + next_dx >= 0 && x + next_dx < lvl->cols && z + next_dz >= 0 && z + next_dz < lvl->rows;
+        int wrap_shimmy = in_bounds && !TEST_BIT(outState->blocks, shimmy_path_cell) && !TEST_BIT(outState->blocks, shimmy_cell);
+        
+        if(basic_shimmy){
+            return shimmy_normal(outState, action, lvl, local_direction, mode, env);
+        }
+        else if(rotation_shimmy){
+            //rotate shimmy
+            static const int LEFT_TURNS[] = {3, 0, 1, 2};   // RIGHT->UP, DOWN->RIGHT, LEFT->DOWN, UP->LEFT
+            static const int RIGHT_TURNS[] = {1, 2, 3, 0};  // RIGHT->DOWN, DOWN->LEFT, LEFT->UP, UP->RIGHT
+
+            outState->robot_orientation = (action == ACTION_LEFT) ? 
+                LEFT_TURNS[outState->robot_orientation] : 
+                RIGHT_TURNS[outState->robot_orientation];
+            outState->robot_state = 1;
+            return 1;
+        }
+        else if(wrap_shimmy){
+            return wrap_around(outState, action, lvl, mode, env);
+        }
+
+    }
+
+    // drop from hang action 
+    if (action == ACTION_DROP && !standing_and_holding_nothing) {
+        return drop_from_hang(outState, action, mode, lvl);
+    }
+    // grab action
+    if (action == ACTION_GRAB && standing_and_holding_nothing 
+    && block_in_front){
+        if (outState->block_grabbed == -1){
+            outState->block_grabbed = next_cell;
+            return 1;
+        } 
+    } 
+    if (action == ACTION_GRAB && outState->block_grabbed != -1){
+        outState->block_grabbed = -1;
+        return 1;
+    }
+    
+    // push or pull block 
+    if (movement_action && block_in_front && outState->block_grabbed != -1){
+        int result = 0;
+        if (outState->robot_orientation == action){
+            result = push(outState, action, lvl, mode, env);
+        } else if(outState->robot_orientation == (action+2)%4){
+            result = pull(outState, action, lvl, mode, env);
+        } else {
+            outState->robot_orientation = action;
+            outState->block_grabbed = -1;
+            result = 1;
+        }
+        // block fell on top of player
+        if (TEST_BIT(outState->blocks, outState->robot_position)){
+            if (mode == PLG_MODE){
+                return MOVE_ILLEGAL;
+            }
+            if (mode == RL_MODE){
+                return MOVE_DEATH;
+            }
+        }
+        return result;
+    }
+    return 0;
+    
+}
+
+
 
 void next_level(CTowerClimb* env){
     if(env->level_number == 2){
@@ -1032,18 +874,10 @@ void next_level(CTowerClimb* env){
         reset(env);
         return;
     }
-    env->rows_cleared= 0;
     env->level_number += 1;
     env->level = levels[env->level_number];
-    env->robot_position = env->level.spawn_location;
-    env->distance_to_goal = get_distance_to_goal(env);
-    env->robot_state = DEFAULT;
-    env->robot_orientation = UP;
-    env->robot_direction = UP;
-    env->block_grabbed = -1;
-    memcpy(env->board_state, env->level.map, env->level.total_length * sizeof(int));
-    memset(env->blocks_to_move, -1, env->level.cols * sizeof(int));
-    memset(env->blocks_to_fall, -1, env->level.total_length * sizeof(int));
+    levelToPuzzleState(&env->level, &env->state);
+
 }
 
 void step(CTowerClimb* env) {
@@ -1055,72 +889,423 @@ void step(CTowerClimb* env) {
 	//        add_log(env->log_buffer, &env->log);
 	//        reset(env);
     // }
-    int action = env->actions[0];
-    if (action == LEFT || action == RIGHT || action == DOWN || action == UP) {
-        int direction = get_direction(env, action);
-        int moving_block = (env->block_grabbed != -1 && abs(env->robot_orientation - action) == 2);
-        if(moving_block){
-            env->rewards[0] = env->reward_move_block;
-            env->log.episode_return += env->reward_move_block;
-        }
-        if (direction == env->robot_orientation || moving_block || env->robot_state == HANGING){
-            env->robot_direction = direction;
-            handle_move_forward(env, action);
-            int sz = env->level.size;
-            int below_index = env->robot_position - sz;
-            // check if goal is below current position
-            if(below_index < 0){
-                env->log.rows_cleared = env->rows_cleared;
-                compute_observations(env);
-                int distance = get_distance_to_goal(env);
-                int delta_distance = env->distance_to_goal - distance;
-                float distance_reward = delta_distance * env->reward_distance;
-                env->rewards[0] += distance_reward;
-                env->log.episode_return += distance_reward;
-	        }
-            else if (env->board_state[below_index] == 2){
-                env->rewards[0] = 1;
-                env->log.episode_return += 1;
-                env->log.rows_cleared = env->rows_cleared;
-                env->log.levels_completed = env->level_number + 1;
-                add_log(env->log_buffer, &env->log);
-                next_level(env);
-                //reset(env);
-            }
-            
-        }
-        else {
-            env->robot_direction = direction;
-            env->robot_orientation = direction;
-        }
+    // Create next state
+    PuzzleState nextState = env->state;
+    int move_result = applyAction(&nextState, env->actions[0], &env->level, RL_MODE, env);
+    
+    if (move_result == MOVE_ILLEGAL) {
+        illegal_move(env);
+        return;
     }
-	
-    else if (action == GRAB){
-        handle_grab_block(env);
+    if (move_result == MOVE_DEATH){
+        death(env);
+        return;
     }
-    else if (action == DROP){
-        env->block_grabbed = -1;
-        if(env->robot_state == HANGING) handle_drop(env);
+    
+    // Update state
+    env->state = nextState;
+    
+    // Check for goal state
+    if (isGoal(&env->state, &env->level)) {
+        next_level(env);
     }
-    // check if block fell on you
-    if(env->board_state[env->robot_position] == 1){
-        env->rewards[0] = -1;
-        env->log.episode_return -= 1;
-        add_log(env->log_buffer, &env->log);
-        reset(env);
-    }
-    env->log.rows_cleared = env->rows_cleared;
-    int distance = get_distance_to_goal(env);
-    int delta_distance = env->distance_to_goal - distance;
-    float distance_reward = delta_distance * env->reward_distance;
-    env->rewards[0] += distance_reward;
-    env->log.episode_return += distance_reward;
-    env->distance_to_goal = distance;
+    
+    // Update observations
     compute_observations(env);
     // if(action != NOOP){
     //     print_observation_window(env);
     // }
 }
+
+
+typedef struct BFSNode {
+    PuzzleState state;
+    int depth;      // how many moves from start
+    int parent;     // index in BFS array of who generated me
+    int action;     // which action led here (if you want to reconstruct the path)
+} BFSNode;
+
+static BFSNode* queueBuffer = NULL;
+static int front = 0, back = 0;
+
+// hash table for visited states
+typedef struct VisitedNode {
+    PuzzleState state;
+    uint64_t hashVal;
+    struct VisitedNode* next;
+} VisitedNode;
+
+static VisitedNode* visitedTable[TABLE_SIZE];
+
+
+// Helper to incorporate a 32-bit integer into the hash one byte at a time.
+static inline uint64_t fnv1a_hash_int(uint64_t h, int value) {
+    // Break the int into 4 bytes (assuming 32-bit int).
+    // This ensures consistent hashing regardless of CPU endianness.
+    unsigned char bytes[4];
+    bytes[0] = (unsigned char)((value >>  0) & 0xFF);
+    bytes[1] = (unsigned char)((value >>  8) & 0xFF);
+    bytes[2] = (unsigned char)((value >> 16) & 0xFF);
+    bytes[3] = (unsigned char)((value >> 24) & 0xFF);
+
+    for (int i = 0; i < 4; i++) {
+        h ^= bytes[i];
+        h *= FNV_PRIME;
+    }
+    return h;
+}
+
+uint64_t hashPuzzleState(const PuzzleState *s) {
+    uint64_t h = FNV_OFFSET;
+
+    // 1) Hash the 125-byte bitmask
+    for (int i = 0; i < BLOCK_BYTES; i++) {
+        h ^= s->blocks[i];
+        h *= FNV_PRIME;
+    }
+
+    // 2) Hash the int fields (position, orientation, state, block_grabbed)
+    h = fnv1a_hash_int(h, s->robot_position);
+    h = fnv1a_hash_int(h, s->robot_orientation);
+    h = fnv1a_hash_int(h, s->robot_state);
+    h = fnv1a_hash_int(h, s->block_grabbed);
+
+    return h;
+}
+// Compares two puzzle states fully
+int equalPuzzleState(const PuzzleState* a, const PuzzleState* b) {
+    // compare bitmask
+    if (memcmp(a->blocks, b->blocks, BLOCK_BYTES) != 0) return 0;
+    // compare other fields
+    if (a->robot_position != b->robot_position) return 0;
+    if (a->robot_orientation != b->robot_orientation) return 0;
+    if (a->robot_state != b->robot_state) return 0;
+    if (a->block_grabbed != b->block_grabbed) return 0;
+    return 1;
+}
+
+void resetVisited(void) {
+    // If we want to free old chains, we do that, or just memset to 0 for now.
+    memset(visitedTable, 0, sizeof(visitedTable));
+}
+
+// Helper function to find a node in the hash table
+static VisitedNode* findNode(const PuzzleState* s, uint64_t hv, size_t idx) {
+    VisitedNode* node = visitedTable[idx];
+    while (node) {
+        if (node->hashVal == hv && equalPuzzleState(&node->state, s)) {
+            return node;
+        }
+        node = node->next;
+    }
+    return NULL;
+}
+
+int isVisited(const PuzzleState* s) {
+    uint64_t hv = hashPuzzleState(s);
+    size_t idx = (size_t)(hv % TABLE_SIZE);
+    return findNode(s, hv, idx) != NULL;
+}
+
+void markVisited(const PuzzleState* s) {
+    uint64_t hv = hashPuzzleState(s);
+    size_t idx = (size_t)(hv % TABLE_SIZE);
+
+    // Return if already present
+    if (findNode(s, hv, idx)) {
+        return;
+    }
+
+    // Insert new node
+    VisitedNode* node = (VisitedNode*)malloc(sizeof(VisitedNode));
+    node->state = *s;
+    node->hashVal = hv;
+    node->next = visitedTable[idx];
+    visitedTable[idx] = node;
+}
+
+void print_level_detailed(const Level* lvl, const unsigned char* bitmask, int robot_position) {
+    printf("\nLevel layout (Height layers from bottom to top):\n");
+    
+    // For each height level (total_length / size gives us height)
+    int height = lvl->total_length / lvl->size;
+    for (int y = 0; y < height; y++) {
+        printf("\nLayer %d:\n", y);
+        
+        // For each depth row
+        for (int z = 0; z < lvl->rows; z++) {
+            printf("  ");  // Indent
+            
+            // For each width position
+            for (int x = 0; x < lvl->cols; x++) {
+                // Calculate index in the same way as the original level layout
+                int idx = x + lvl->cols * z + lvl->size * y;
+                
+                if (idx == lvl->spawn_location) {
+                    printf("S ");
+                    continue;
+                }      
+                if (idx == robot_position) {
+                    printf("R ");
+                    continue;
+                }
+                // Print appropriate symbol based on bitmask test
+                if (idx == lvl->goal_location) {
+                    printf("2 ");  // Goal location
+                } else if (TEST_BIT(bitmask, idx)) {
+                    printf("■ ");  // Block present
+                } else {
+                    printf("□ ");  // Empty space
+                }
+            }
+            printf("\n");
+        }
+    }
+    printf("\n");
+}
+
+void print_state(const PuzzleState* s, const Level* lvl){
+    print_level_detailed(lvl, s->blocks, s->robot_position);
+    printf("robot position: %d\n", s->robot_position);
+    printf("robot orientation: %d\n", s->robot_orientation);
+    printf("robot state: %d\n", s->robot_state);
+    printf("block grabbed: %d\n", s->block_grabbed);
+}
+
+// This function fills out up to MAX_NEIGHBORS BFSNodes in 'outNeighbors'
+// from a given BFSNode 'current'. It returns how many neighbors it produced.
+int getNeighbors(const BFSNode* current, BFSNode* outNeighbors, const Level* lvl) {
+    int count = 0;
+    // We'll read the current BFSNode's puzzle state
+    const PuzzleState* curState = &current->state;
+
+    // Try each action
+    for (int i = 0; i < 6; i++) {
+        int action = i;
+
+        // 1) Make a copy of the current puzzle state
+        PuzzleState newState = *curState;
+
+        // 2) Attempt to apply the action to newState
+        int success = applyAction(&newState, action, lvl, PLG_MODE, NULL);
+
+        if (!success) {
+            // Move was invalid, skip
+            continue;
+        }
+
+        // 3) If valid, build a BFSNode
+        BFSNode neighbor;
+        neighbor.state = newState;
+        neighbor.depth = current->depth + 1;
+        neighbor.parent = -1;   // BFS sets or overwrites this later
+        neighbor.action = action; // record which action led here
+
+        // 4) Add to 'outNeighbors' array
+        outNeighbors[count++] = neighbor;
+
+        // If you only allow up to 6 total, we can break if we reach that
+        if (count >= MAX_NEIGHBORS) break;
+    }
+
+    return count; // how many valid neighbors we produced
+}
+
+// Example BFS
+int bfs(const PuzzleState* start, int maxDepth, const Level* lvl) {
+    // Clear or init your BFS queue
+    queueBuffer = (BFSNode*)malloc(MAX_BFS_SIZE * sizeof(BFSNode));
+    if (!queueBuffer) {
+        printf("Failed to allocate memory for BFS queue\n");
+        return 0;
+    }
+    
+    front = 0;
+    back = 0;
+    
+    // Enqueue start node
+    BFSNode startNode;
+    startNode.state = *start;  // copy puzzle state
+    startNode.depth = 0;
+    startNode.parent = -1;
+    startNode.action = -1;
+    queueBuffer[back++] = startNode;
+
+    // BFS loop
+    while (front < back) {
+
+        if (back >= MAX_BFS_SIZE) {
+            printf("BFS queue overflow! Increase MAX_BFS_SIZE or optimize search.\n");
+            return 0;
+        }
+        BFSNode current = queueBuffer[front];
+        int currentIndex = front;
+        front++;
+        // If current.state is the goal, reconstruct path
+        if (isGoal(&current.state, lvl)) {
+            printf("Found solution path of length %d!\n", current.depth);
+            
+            // Store nodes in order
+            BFSNode* path = (BFSNode*)malloc((current.depth + 1) * sizeof(BFSNode));
+            BFSNode node = current;
+            int idx = current.depth;
+            
+            // Walk backwards to get path
+            while (idx >= 0) {
+                path[idx] = node;
+                if (node.parent != -1) {
+                    node = queueBuffer[node.parent];
+                }
+                idx--;
+            }
+            
+            // Print in forward order
+            printf("\nStep 0 (Start):\n");
+            printf("  Position: %d\n", path[0].state.robot_position);
+            printf("  Orientation: %d\n", path[0].state.robot_orientation);
+            printf("  State: %d\n", path[0].state.robot_state);
+            printf("  Block grabbed: %d\n", path[0].state.block_grabbed);
+            for (int i = 1; i <= current.depth; i++) {
+                int block_in_front = path[i].state.robot_position + 
+                    (path[i].state.robot_orientation == 0 ? 1 :    // Right
+                     path[i].state.robot_orientation == 1 ? lvl->cols :    // Down
+                     path[i].state.robot_orientation == 2 ? -1 :   // Left
+                     -lvl->cols);  // Up (orientation == 3)
+                printf("\nStep %d:\n", i);
+                printf("  Action taken: %d\n", path[i].action);
+                printf("  Position: %d\n", path[i].state.robot_position);
+                printf("  Orientation: %d\n", path[i].state.robot_orientation);
+                printf("  State: %d\n", path[i].state.robot_state);
+                printf("  Block grabbed: %d\n", path[i].state.block_grabbed);
+                printf("  Block in front: %d (Has block: %d)\n", block_in_front, 
+                    bfs_is_valid_position(block_in_front, lvl) && 
+                    (TEST_BIT(path[i].state.blocks, block_in_front) || block_in_front == lvl->goal_location));
+            }
+            
+            free(path);
+            return 1;
+        }
+
+        if (current.depth < maxDepth) {
+            // generate neighbors
+            BFSNode neighbors[MAX_NEIGHBORS];
+            int nCount = getNeighbors(&current, neighbors, lvl);
+            for (int i = 0; i < nCount; i++) {
+                PuzzleState* nxt = &neighbors[i].state;
+                // if not visited
+                if (!isVisited(nxt)) {
+                    markVisited(nxt);
+
+                    // fill BFS node fields
+                    neighbors[i].depth = current.depth + 1;
+                    neighbors[i].parent = currentIndex;
+                    // Enqueue
+                    queueBuffer[back++] = neighbors[i];
+                }
+            }
+        }        
+    }
+    free(queueBuffer);
+    queueBuffer = NULL;
+    // If we exit while, no solution found within maxDepth
+    printf("No solution within %d moves.\n", maxDepth);
+    return 0;
+}
+
+
+int verify_level(Level level, int max_moves){
+    // converting level to puzzle state
+    PuzzleState state;
+    levelToPuzzleState(&level, &state);
+    // reset visited hash table
+    resetVisited();
+    markVisited(&state);
+    // Run BFS
+    int solvable = bfs(&state, max_moves, &level);
+    return solvable;
+}
+
+Level gen_level(int goal_level) {
+    // Initialize an illegal level in case we need to return early
+    Level illegal_level = {0};
+    
+    // Allocate board memory
+    int* board = (int*)calloc(row_max * col_max * depth_max, sizeof(int));
+    if (!board) {
+        return illegal_level;
+    }
+
+    int legal_width_size = 8;
+    int legal_depth_size = 8;
+    int area = depth_max * col_max;
+    int spawn_created = 0;
+    int spawn_index = -1;
+    int goal_created =0;
+    int goal_index = -1;
+    for(int y= 0; y < row_max; y++){
+        for(int z = depth_max - 1; z >= 0; z--){
+            for(int x = 0; x< col_max; x++){
+                int block_index = x + col_max * z + area * y;
+                if (x >= 1 && x < legal_width_size && z >= 1 && z < legal_depth_size && 
+                y >= 1 && y < goal_level && (z <= (legal_depth_size - y))){
+                    int chance = (rand() % 4 ==0) ? 1 : 0;
+                    board[block_index] = chance;
+                    // create spawn point above an existing block
+                    if (spawn_created == 0 && y == 2 && board[block_index - area] == 1){
+                        spawn_created = 1;
+                        spawn_index = block_index;
+                        board[spawn_index] = 0;
+                    }
+                }
+                if (goal_created ==0 && y == goal_level && (board[block_index + col_max  - area]  ==1 ||  board[block_index - 1  - area] == 1|| board[block_index + 1 - area] ==1 )){
+                    goal_created = 1;
+                    goal_index = block_index;
+                    board[goal_index] = 2;
+                }
+                
+            }
+        }
+    }
+
+    if (!spawn_created || spawn_index < 0) {
+        printf("no spawn found\n");
+        free(board);
+        return illegal_level;
+    }
+
+    
+    if (!goal_created || goal_index < 0) {
+        printf("no goal found\n");
+        printf("goal index: %d\n", goal_index);
+        free(board);
+        return illegal_level;
+    }
+
+    Level level = {
+        .map = board,
+        .rows = row_max,
+        .cols = col_max,
+        .size = row_max * col_max,
+        .total_length = row_max * col_max * depth_max,
+        .goal_location = goal_index,
+        .spawn_location = spawn_index
+    };
+    return level;
+}
+
+
+static void init_random_levels(int goal_level) {
+    time_t t;
+    for(int i = 0; i < 3; i++) {
+        srand((unsigned) time(&t) + i); // Increment seed for each level
+        levels[i] = gen_level(goal_level);
+        // guarantee a map is created
+        while(levels[i].map == NULL || verify_level(levels[i],32) == 0){
+            levels[i] = gen_level(goal_level);
+        }
+    }
+}
+
 
 const Color STONE_GRAY = (Color){80, 80, 80, 255};
 const Color PUFF_RED = (Color){187, 0, 0, 255};
@@ -1255,11 +1440,11 @@ Client* make_client(CTowerClimb* env) {
         client->cube.materials[i].shader = client->shader;
     }
     client->animState = ANIM_IDLE;
-    client->previousRobotPosition = env->robot_position;
+    client->previousRobotPosition = env->state.robot_position;
     
     // Initialize visual position to match starting robot position
-    int floor = env->robot_position / env->level.size;
-    int grid_pos = env->robot_position % env->level.size;
+    int floor = env->state.robot_position / env->level.size;
+    int grid_pos = env->state.robot_position % env->level.size;
     int x = grid_pos % env->level.cols;
     int z = grid_pos / env->level.cols;
     
@@ -1276,17 +1461,16 @@ Client* make_client(CTowerClimb* env) {
 void orient_hang_offset(Client* client, CTowerClimb* env, int reverse){
 
     client->visualPosition.y -= 0.2f * reverse;
-    if (env->robot_orientation == 0) { // Facing +x
+    if (env->state.robot_orientation == 0) { // Facing +x
         client->visualPosition.x += 0.4f * reverse;
-    } else if (env->robot_orientation == 1) { // Facing +z
+    } else if (env->state.robot_orientation == 1) { // Facing +z
         client->visualPosition.z += 0.4f * reverse;
-    } else if (env->robot_orientation == 2) { // Facing -x
+    } else if (env->state.robot_orientation == 2) { // Facing -x
         client->visualPosition.x -= 0.4f * reverse;
-    } else if (env->robot_orientation == 3) { // Facing -z
+    } else if (env->state.robot_orientation == 3) { // Facing -z
         client->visualPosition.z -= 0.4f * reverse;
     }
 }
-
 
 void render(Client* client, CTowerClimb* env) {
     if (IsKeyDown(KEY_ESCAPE)) {
@@ -1295,12 +1479,12 @@ void render(Client* client, CTowerClimb* env) {
     static int stored_orientation = -1;  // Store original orientation during climb
     int cols = env->level.cols;
     int sz = env->level.size;
-    int floor = env->robot_position / sz;
-    int grid_pos = env->robot_position % sz;
+    int floor = env->state.robot_position / sz;
+    int grid_pos = env->state.robot_position % sz;
     int x = grid_pos % cols;
     int z = grid_pos / cols;
 
-    if (env->robot_state == DEFAULT && client->animState == ANIM_HANGING) {
+    if (env->state.robot_state == DEFAULT && client->animState == ANIM_HANGING) {
         client->animState = ANIM_IDLE;
         client->animFrameCounter = 0;
         UpdateModelAnimation(client->robot, client->animations[4], 0);
@@ -1308,7 +1492,7 @@ void render(Client* client, CTowerClimb* env) {
         client->visualPosition = client->targetPosition;
     }
 
-    if (env->block_grabbed != -1 && client->animState != ANIM_GRABBING && client->animState != ANIM_START_GRABBING)
+    if (env->state.block_grabbed != -1 && client->animState != ANIM_GRABBING && client->animState != ANIM_START_GRABBING)
     {
         client->animState = ANIM_START_GRABBING;
         client->isMoving = true;
@@ -1316,7 +1500,7 @@ void render(Client* client, CTowerClimb* env) {
         UpdateModelAnimation(client->robot, client->animations[3], 0); 
         client->visualPosition = client->targetPosition;
 
-    } else if (env->block_grabbed == -1 && client->animState == ANIM_GRABBING) {
+    } else if (env->state.block_grabbed == -1 && client->animState == ANIM_GRABBING) {
         // Reset to idle when block is released
         client->animState = ANIM_IDLE;
         client->animFrameCounter = 0;
@@ -1324,7 +1508,7 @@ void render(Client* client, CTowerClimb* env) {
     }
 
     // Check if robot position has changed
-    if (env->robot_position != client->previousRobotPosition) {
+    if (env->state.robot_position != client->previousRobotPosition) {
         if (client->isMoving) {
             client->visualPosition = client->targetPosition;
         }
@@ -1332,8 +1516,8 @@ void render(Client* client, CTowerClimb* env) {
         client->isMoving = true;
             
         // Calculate new target position
-        int floor = env->robot_position / env->level.size;
-        int grid_pos = env->robot_position % env->level.size;
+        int floor = env->state.robot_position / env->level.size;
+        int grid_pos = env->state.robot_position % env->level.size;
         int x = grid_pos % env->level.cols;
         int z = grid_pos / env->level.cols;
         
@@ -1354,7 +1538,7 @@ void render(Client* client, CTowerClimb* env) {
             client->animState = ANIM_CLIMBING;  // Need to add this to AnimationState enum
             UpdateModelAnimation(client->robot, client->animations[1], 0);  // Assuming climbing is animation 4
             client->animFrameCounter = 6;
-        } else if (env->robot_state == HANGING){
+        } else if (env->state.robot_state == HANGING){
             client->isMoving = true;
             // Check for wrap shimmy case - when both x and z change
             bool is_wrap_shimmy = fabs(client->targetPosition.x - client->visualPosition.x) > 0.5f && 
@@ -1371,13 +1555,13 @@ void render(Client* client, CTowerClimb* env) {
             } else {
                 // Regular shimmy animation logic
                 bool moving_right = false;
-                if (env->robot_orientation == UP) {  // Facing up (-z)
+                if (env->state.robot_orientation == UP) {  // Facing up (-z)
                     moving_right = client->targetPosition.x > client->visualPosition.x;
-                } else if (env->robot_orientation == DOWN) {  // Facing down (+z)
+                } else if (env->state.robot_orientation == DOWN) {  // Facing down (+z)
                     moving_right = client->targetPosition.x < client->visualPosition.x;
-                } else if (env->robot_orientation == RIGHT) {  // Facing right (+x)
+                } else if (env->state.robot_orientation == RIGHT) {  // Facing right (+x)
                     moving_right = client->targetPosition.z < client->visualPosition.z;
-                } else if (env->robot_orientation == LEFT) {  // Facing left (-x)
+                } else if (env->state.robot_orientation == LEFT) {  // Facing left (-x)
                     moving_right = client->targetPosition.z > client->visualPosition.z;
                 }
                 
@@ -1412,7 +1596,7 @@ void render(Client* client, CTowerClimb* env) {
             }
             
         }
-        client->previousRobotPosition = env->robot_position;
+        client->previousRobotPosition = env->state.robot_position;
     }
 
     // Update animation if moving
@@ -1420,10 +1604,10 @@ void render(Client* client, CTowerClimb* env) {
         // Use appropriate animation based on state
         if (client->animState == ANIM_CLIMBING) {
             if (stored_orientation == -1) {
-                stored_orientation = env->robot_orientation;
+                stored_orientation = env->state.robot_orientation;
             }
-            int current_orientation = env->robot_orientation;
-            env->robot_orientation = stored_orientation;
+            int current_orientation = env->state.robot_orientation;
+            env->state.robot_orientation = stored_orientation;
             client->animFrameCounter += 6;
             UpdateModelAnimation(client->robot, client->animations[1], client->animFrameCounter);
             if (client->animFrameCounter >= client->animations[1].frameCount) {  // Adjust frame count for climbing animation
@@ -1433,7 +1617,7 @@ void render(Client* client, CTowerClimb* env) {
                 UpdateModelAnimation(client->robot, client->animations[4], client->animFrameCounter);
                 client->visualPosition = client->targetPosition;
                 stored_orientation = -1;
-                env->robot_orientation = current_orientation;
+                env->state.robot_orientation = current_orientation;
             }
         } else if (client->animState == ANIM_HANGING){
             client->animFrameCounter = 0;
@@ -1452,13 +1636,13 @@ void render(Client* client, CTowerClimb* env) {
             if (progress > 1.0f) progress = 1.0f;
             
             // Lerp position based on orientation
-            if (env->robot_orientation == UP) {  // Facing -z
+            if (env->state.robot_orientation == UP) {  // Facing -z
                 client->visualPosition.x = Lerp(client->visualPosition.x, client->targetPosition.x, progress);
-            } else if (env->robot_orientation == DOWN) {  // Facing +z
+            } else if (env->state.robot_orientation == DOWN) {  // Facing +z
                 client->visualPosition.x = Lerp(client->visualPosition.x, client->targetPosition.x, progress);
-            } else if (env->robot_orientation == RIGHT) {  // Facing +x
+            } else if (env->state.robot_orientation == RIGHT) {  // Facing +x
                 client->visualPosition.z = Lerp(client->visualPosition.z, client->targetPosition.z, progress);
-            } else if (env->robot_orientation == LEFT) {  // Facing -x
+            } else if (env->state.robot_orientation == LEFT) {  // Facing -x
                 client->visualPosition.z = Lerp(client->visualPosition.z, client->targetPosition.z, progress);
             }
             if (client->animFrameCounter >= 87) {
@@ -1559,25 +1743,15 @@ void render(Client* client, CTowerClimb* env) {
     BeginBlendMode(BLEND_ALPHA);
     int total_length = env->level.total_length;
     for(int i= 0; i < total_length; i++){
-        if(env->board_state[i] > 0){
-            int floor = i / sz;
-            int grid_pos = i % sz;
-            int x = grid_pos % cols;
-            int z = grid_pos / cols;
-            if (env->board_state[i] == 2){
-                EndShaderMode();
-                DrawCube(
-                (Vector3){x * 1.0f, floor * 1.0f, z * 1.0f},
-                1.0f, 1.0f, 1.0f,
-                PUFF_CYAN
-                );
-                BeginShaderMode(client->shader);
-
-            } else {
-                DrawModel(client->cube, (Vector3){x * 1.0f, floor * 1.0f, z * 1.0f}, client->scale, WHITE);
-            }
+        int floor = i / sz;
+        int grid_pos = i % sz;
+        int x = grid_pos % cols;
+        int z = grid_pos / cols;
+        if(TEST_BIT(env->state.blocks, i)){
             
-            if(i == env->block_grabbed){
+                DrawModel(client->cube, (Vector3){x * 1.0f, floor * 1.0f, z * 1.0f}, client->scale, WHITE);
+            
+            if(i == env->state.block_grabbed){
                 DrawCubeWires(
                 (Vector3){x * 1.0f, floor * 1.0f, z * 1.0f},
                 1.0f, 1.0f, 1.0f,
@@ -1590,6 +1764,15 @@ void render(Client* client, CTowerClimb* env) {
             );
             }
             
+        }
+        if (i == env->level.goal_location){
+            EndShaderMode();
+            DrawCube(
+            (Vector3){x * 1.0f, floor * 1.0f, z * 1.0f},
+            1.0f, 1.0f, 1.0f,
+            PUFF_CYAN
+            );
+            BeginShaderMode(client->shader);
         }
     }
     EndBlendMode();
@@ -1608,7 +1791,7 @@ void render(Client* client, CTowerClimb* env) {
     rlPushMatrix();
     rlTranslatef(spherePos.x, spherePos.y, spherePos.z);
     rlRotatef(90.0f, 1, 0, 0);
-    rlRotatef(-90.0f + env->robot_orientation * 90.0f, 0, 0, 1);
+    rlRotatef(-90.0f + env->state.robot_orientation * 90.0f, 0, 0, 1);
     DrawModel(client->robot, (Vector3){ 0, 0, 0}, 0.5f, WHITE);
     rlPopMatrix();
